Create GrowBox.py
The main idea was to create a grow box for internal plant growing with adjustable claimant parameters (light in this case). For such  purposes were used: photocell sensor, capacitor, dc motor, breadboard with some wires, Raspberry pi and a little python. So photocell sensor is connected with capacitor at the breadboard. By time needed to charge the capacitor we can define solar insulation level in order to compare it with optimal range for plant growing. If received data is bigger or smaller then extreme values of solar insulating, dc motor (connected to a jalousie) opens/closes (change the angle of) jalousie until photocell gets optimal amount of light. In perfect scenario humidity sensor, temperature sensor could be used along with humidifier, infrared lamp and ventilator in order to create optimal micro-climate in grow box, but because of luck of equipment and impossibility to test the code I've just left it on paper.